refactor: invert rules db dependency
To retrieve a best_practices and code_smells is required to do an inner
join with repository table. This is a performance issue, because a lot
of data would be retrieved per query.

Instead of each table having a repository FK create a new table called
project_rules which has the FK of each rules table and the project (aka
repository).

- Rename repository table to project.
- Update RulesRepo to request a repository id as parameter to create a
  project_rules entry.
- Remove repository FK from rule tables.
- Update RulesRepo createRule() method to return an ID of the created
  ProjectRules entry.