feat: request project id on rules creation
The system is creating a new project per POST request, but the endpoint
responsibility is to only create the project rules.

Update system to ask for the project ID as a parameter of the endpoint,
this way it will validate the project ID exists and does not have
existent defined rules.

- Update handler endpoint to receive project parameter.
- Update service to verify if the project exists and if it has rules
  already.